Moving namespaces tests
